Refactor: Home Directory rename 및 구조 변경
