+++ src/components/Common/CandidateCard/index.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import styled, { css } from 'styled-components';
 import theme from '../../../lib/styles/theme';
 import media from '../../../lib/styles/media';
-import CardItem from '../../Common/CandidateCard/CardItem';
+import CardItem from './CardItem';
 
 type Runner = {
   id: number;
