+++ src/components/Home/Classification/Candidate/EmptyCard.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import styled from 'styled-components';
-import theme from '../../../lib/styles/theme';
-import media from '../../../lib/styles/media';
+import theme from '../../../../lib/styles/theme';
+import media from '../../../../lib/styles/media';
 
 type EmptyCardProps = {
   url: string;
