+++ src/components/Home/Classification/Category/CategoryItem.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import styled, { css } from 'styled-components';
-import theme from '../../../lib/styles/theme';
-import media from '../../../lib/styles/media';
+import theme from '../../../../lib/styles/theme';
+import media from '../../../../lib/styles/media';
 
 type CategoryItemProps = {
   title: string;
