+++ src/components/Home/HomeBoard/InfoCard/Slide.tsx
