+++ src/components/Home/HomeBoard/Notice/NoticeArticle.tsx
@@ -1,8 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import styled from 'styled-components';
-import bannerData from '../../../api/BannerInfo.json';
-// import bannerIcon from '../../../../public/img/bannerIcon.svg';
-import media from '../../../lib/styles/media';
+import bannerData from '../../../../api/BannerInfo.json';
+import media from '../../../../lib/styles/media';
 
 type dataType = {
   id: number,
